worker/requirements.txt
kafka-python
pymysql
nltk
python-dotenv
